requirements.txt
flask>=3.1.2
pyyaml>=6.0.3
